+++ lotus/admin/article.py
@@ -117,7 +117,9 @@ class ArticleAdmin(SmartModelAdmin):
         """
         Return humanized name for object language code.
         """
-        return LANGUAGE_NAMES[obj.language]
+        if obj.language in LANGUAGE_NAMES:
+            return LANGUAGE_NAMES[obj.language]
+        return "{} (disabled)".format(obj.language)
     language_name.short_description = _("language")
     language_name.admin_order_field = "language"